test: put host binding acceptances tests in their own file (#29946)
PR Close #29946
